Refact: 리뷰 및 질문 관련 DTO에서 memberId 필드 제거, 삭제 요청 DTO 삭제, 인증 정보 기반으로 리뷰 및 질문 생성/수정/삭제 메소드 수정

+++ src/main/java/com/shopping/service/QuestionService.java
@@ -4,10 +4,10 @@ import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
 import com.shopping.dto.Request.QuestionCreateRequestDto;
-import com.shopping.dto.Request.QuestionDeleteRequestDto;
 import com.shopping.dto.Request.QuestionUpdateRequestDto;
 import com.shopping.dto.Response.AnswerResponseDto;
 import com.shopping.dto.Response.QuestionResponseDto;
@@ -16,6 +16,7 @@ import com.shopping.model.Member;
 import com.shopping.model.Question;
 import com.shopping.repository.MemberRepository;
 import com.shopping.repository.QuestionRepository;
+import com.shopping.util.SecurityUtil;
 
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
@@ -29,8 +30,8 @@ public class QuestionService {
     private final MemberRepository memberRepository;
     private final QuestionRepository questionRepository;
 
-    public QuestionResponseDto createQuestion(QuestionCreateRequestDto requestDto) {
-        Member member = memberRepository.findById(requestDto.getMemberId())
+    public QuestionResponseDto createQuestion(QuestionCreateRequestDto requestDto, Long memberId) {
+        Member member = memberRepository.findById(memberId)
                 .orElseThrow(() -> new IllegalArgumentException("해당 회원이 존재하지 않습니다."));
         Question question = requestDto.toEntity(member);
         questionRepository.save(question);
@@ -46,16 +47,21 @@ public class QuestionService {
         return responseDtos;
     }
 
-    public QuestionResponseDto getQuestion(Long id){
+    public QuestionResponseDto getQuestion(Long id) {
         Question question = questionRepository.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("해당 질문이 존재하지 않습니다."));
         QuestionResponseDto responseDto = QuestionResponseDto.fromEntity(question);
         return responseDto;
     }
 
-    public QuestionResponseDto updateQuestion(QuestionUpdateRequestDto requestDto){
-        Question question = questionRepository.findById(requestDto.getQuestionId())
+    public QuestionResponseDto updateQuestion(QuestionUpdateRequestDto requestDto, Long memberId, Long questionId) {
+        Question question = questionRepository.findById(questionId)
                 .orElseThrow(() -> new IllegalArgumentException("해당 질문이 존재하지 않습니다."));
+
+        if (!SecurityUtil.isAdminOrOwner(question.getMember().getId(), memberId)) {
+            throw new AccessDeniedException("해당 질문을 업데이트할 권한이 없습니다.");
+        }
+
         question.setContent(requestDto.getContent());
         question.setTitle(requestDto.getTitle());
         question.setUpDateTime(LocalDateTime.now());
@@ -65,17 +71,21 @@ public class QuestionService {
     }
 
     @Transactional
-    public void deleteQuestion(Long questionId,QuestionDeleteRequestDto requestDto){
+    public void deleteQuestion(Long questionId, Long memberId) {
         Question question = questionRepository.findById(questionId)
                 .orElseThrow(() -> new IllegalArgumentException("해당 질문이 존재하지 않습니다."));
+        if (!SecurityUtil.isAdminOrOwner(question.getMember().getId(), memberId)) {
+            throw new AccessDeniedException("해당 질문을 삭제할 권한이 없습니다.");
+        }
+
         questionRepository.delete(question);
     }
 
-    public List<AnswerResponseDto> getAnswers(Long questionId){
+    public List<AnswerResponseDto> getAnswers(Long questionId) {
         Question question = questionRepository.findById(questionId)
                 .orElseThrow(() -> new IllegalArgumentException("해당 질문이 존재하지 않습니다."));
-        
-        List<Answer> answers=question.getAnswers();
+
+        List<Answer> answers = question.getAnswers();
         List<AnswerResponseDto> responseDto = answers.stream()
                 .map(AnswerResponseDto::fromEntity)
                 .collect(Collectors.toList());
